requirement.txt
async-timeout            4.0.2
autopep8                 1.6.0
bidict                   0.22.0
boto3                    1.24.4
botocore                 1.27.4
CacheControl             0.12.11
cachetools               5.2.0
certifi                  2022.5.18.1
charset-normalizer       2.0.12
click                    8.1.3
Deprecated               1.2.13
dnspython                2.2.1
elastic-transport        8.1.2
elasticsearch            8.2.2
eventlet                 0.33.1
firebase-admin           5.2.0
Flask                    2.1.2
Flask-SocketIO           5.2.0
google-api-core          2.8.1
google-api-python-client 2.50.0
google-auth              2.7.0
google-auth-httplib2     0.1.0
google-cloud-core        2.3.0
google-cloud-firestore   2.5.2
google-cloud-storage     2.3.0
google-crc32c            1.3.0
google-resumable-media   2.3.3
googleapis-common-protos 1.56.2
greenlet                 1.1.2
grpcio                   1.46.3
grpcio-status            1.46.3
httplib2                 0.20.4
idna                     3.3
importlib-metadata       4.11.4
itsdangerous             2.1.2
Jinja2                   3.1.2
jmespath                 1.0.0
MarkupSafe               2.1.1
msgpack                  1.0.4
mysql-connector-python   8.0.29
packaging                21.3
pip                      22.1.2
proto-plus               1.20.5
protobuf                 3.20.1
pyasn1                   0.4.8
pyasn1-modules           0.2.8
pycodestyle              2.8.0
PyJWT                    2.4.0
pyparsing                3.0.9
python-dateutil          2.8.2
python-engineio          4.3.2
python-socketio          5.6.0
pytz                     2022.1
PyYAML                   6.0
redis                    4.3.3
requests                 2.27.1
rsa                      4.8
s3transfer               0.6.0
setuptools               58.1.0
six                      1.16.0
toml                     0.10.2
uritemplate              4.1.1
urllib3                  1.26.9
Werkzeug                 2.1.2
wheel                    0.37.1
wrapt                    1.14.1
zipp                     3.8.0